split out form fields
