+++ packages/web/app/features/ProgramRegistry/index.js
@@ -1,5 +1,5 @@
 export { ProgramRegistryConditionCategoryField } from './ProgramRegistryConditionCategoryField';
 export { ProgramRegistryConditionField } from './ProgramRegistryConditionField';
 export { UpdateConditionFormModal } from './UpdateConditionFormModal';
-export { PatientProgramRegistryUpdateFormModal } from './PatientProgramRegistryUpdateFormModal';
+export { PatientProgramRegistryUpdateModal } from './PatientProgramRegistryUpdateModal';
 export { PatientProgramRegistryActivateModal } from './PatientProgramRegistryActivateModal';
